refac(AgeFarePolicy.java): BigDecimal -> Integer 변경

+++ src/main/java/nextstep/subway/path/domain/Fare.java
@@ -38,7 +38,7 @@ public class Fare {
     }
 
     private void calculateAgePolicy(int age) {
-        this.fare = AgeFarePolicy.calculate(this.fare, age).intValue();
+        this.fare = AgeFarePolicy.calculate(this.fare, age);
     }
 
     public int getFare() {
